+++ prisma/schema.prisma
@@ -29,7 +29,7 @@ model Post {
   content   String
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
-  like      BigInt   @default(0)
+  like      Int   @default(0)
 
   Comment Comment[]
 
@@ -43,7 +43,7 @@ model Comment {
   content   String
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
-  like      BigInt   @default(0)
+  like      Int   @default(0)
 
   User User @relation(fields: [userId], references: [userId], onDelete: Cascade)
   Post Post @relation(fields: [postId], references: [postId], onDelete: Cascade)
